refactored tooltip directive

+++ web/jashboard/directives/TooltipDirective.js
@@ -1,17 +1,17 @@
 (function(module) {
   jashboard.angular = _.extend(module, {
     tooltipDirective: function (tooltipService) {
-      return new jashboard.angular.EventDirectiveDefinition("jbTooltip", function(scope, element, attrs) {
-        var targetSelector = scope.$eval(attrs['jbTooltipFor']);
-        return {
-          show: function() {
+      return function(scope, element, attrs) {
+        var targetSelector = scope.$eval(attrs.jbTooltip);
+        var toggleExpression = attrs.jbTooltipToggle;
+        scope.$watch(toggleExpression, function(newValue, oldValue) {
+          if (newValue) {
             tooltipService.attachHtmlTooltip(targetSelector, element);
-          },
-          hide: function() {
+          } else {
             tooltipService.removeTooltip(targetSelector);
           }
-        };
-      });
+        });
+      };
     }
   });
 }(jashboard.angular || {}));
